Move `Wait` package
