+++ src/common/blockchain/index.js
@@ -3,6 +3,15 @@ import * as Plasma from './plasma'
 import * as Contract from './contract'
 import * as TxOptions from './txOptions'
 import * as Token from './token'
+import * as Wait from './wait'
 import * as BlockchainDataFormatter from './blockchainDataFormatter'
 
-export { Ethereum, Plasma, Contract, TxOptions, BlockchainDataFormatter, Token }
+export {
+  Ethereum,
+  Plasma,
+  Contract,
+  TxOptions,
+  BlockchainDataFormatter,
+  Token,
+  Wait
+}
