+++ src/common/utils/index.js
@@ -18,7 +18,6 @@ import * as Address from './address'
 import * as Validator from './validator'
 import * as Dimensions from './dimensions'
 import * as Reanimated from './reanimated'
-import * as Wait from './wait'
 
 export {
   Address,
@@ -40,6 +39,5 @@ export {
   Transaction,
   Token,
   Validator,
-  Reanimated,
-  Wait
+  Reanimated
 }
